// src/store/modules/app/action-types.ts
/*
 * @Description: app actions type
 * @Author: ZY
 * @Date: 2020-12-23 10:25:37
 * @LastEditors: ZY
 * @LastEditTime: 2020-12-23 14:02:25
 */

export enum AppActionTypes {
  ACTION_TOGGLE_SIDEBAR = 'ACTION_TOGGLE_SIDEBAR',
  ACTION_CLOSE_SIDEBAR = 'ACTION_CLOSE_SIDEBAR',
  ACTION_TOGGLE_DEVICE = 'ACTION_TOGGLE_DEVICE',
  ACTION_SET_SIZE = 'ACTION_SET_SIZE',
  ACTION_QUERY_TYPE = 'ACTION_QUERY_TYPE',
  ACTION_STAT_TYPE = 'ACTION_STAT_TYPE',
}